+++ lib/screens/mobile/mobile.dart
@@ -1,1 +1,1 @@
-export 'dashboard.dart';
+export 'home.dart';
